refactor(nextcloud): extract shared task listener trait and bump to 0.3.33
Review follow-up for #419:

- Move isAiquilaTask() into a shared TaskListenerTrait as a private
  instance method on the injected task processing manager, removing the
  cross-class static reach from TaskFailedListener.
- Move $typeLabels/getTaskTypeLabel() into the trait, which also drops
  the duplicated, unused $typeLabels copy in TaskFailedListener.
- Document that getPreferredProvider() answers "who would run this task
  type now", not "who ran this task": the admin preference is cached, so
  a change between scheduling and completion can misattribute. For
  aiquila's synchronous providers the event fires in-process right after
  the preferred provider ran, so it is accurate in practice.
- Bump the app version to 0.3.33 in info.xml and package.json.
