+++ nextcloud-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aiquila",
-  "version": "0.3.32",
+  "version": "0.3.33",
   "description": "Claude AI Integration for Nextcloud",
   "type": "module",
   "main": "src/main.js",
